fixes after review. CustomAppValidationException moved.
